migrate the code for reading folders and files to api.go
